@@ -11,22 +11,23 @@
     "core-js": "^3.6.5",
     "moment": "^2.29.1",
     "swiper": "^6.4.10",
-    "vue": "^2.6.11",
+    "vue": "^3.0.5",
     "vue-awesome-swiper": "^4.1.1",
     "vue-material": "^1.0.0-beta-15",
     "vuex": "^3.4.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
+    "@babel/core": "^7.12.10",
+    "@babel/eslint-parser": "^7.12.1",
+    "@vue/cli-plugin-babel": "^5.0.0-alpha.3",
+    "@vue/cli-plugin-eslint": "^5.0.0-alpha.3",
+    "@vue/cli-plugin-vuex": "^5.0.0-alpha.3",
+    "@vue/cli-service": "^5.0.0-alpha.3",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^6.2.2",
-    "prettier": "^1.19.1",
+    "eslint": "^7.15.0",
+    "eslint-plugin-prettier": "^3.2.0",
+    "eslint-plugin-vue": "^7.2.0",
+    "prettier": "^2.2.1",
     "vue-template-compiler": "^2.6.11"
   }
 }
