@@ -1,10 +1,12 @@
-import Vue from "vue";
+// import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
-import moment from "moment";
+// import moment from "moment";
 
-Vue.config.productionTip = false;
-Vue.prototype.$moment = moment;
+// Vue.config.productionTip = false;
+// Vue.prototype.$moment = moment;
 
-new Vue({
-  render: (h) => h(App)
-}).$mount("#app");
+createApp(App).mount("#app");
+// new Vue({
+//   render: (h) => h(App)
+// }).$mount("#app");
