@@ -1,5 +1,10 @@
 <template>
-  <swiper ref="mySwiper" :options="swiperOptions">
+  <swiper
+    :slides-per-view="3"
+    :space-between="50"
+    @swiper="onSwiper"
+    @slideChange="onSlideChange"
+  >
     <swiper-slide>
       <DayCardHandler :cardsData="daysData['05.06.2021']" />
     </swiper-slide>
@@ -14,7 +19,7 @@
 </template>
 
 <script>
-import { Swiper, SwiperSlide } from "vue-awesome-swiper";
+import { Swiper, SwiperSlide } from "swiper/vue";
 import DayCardHandler from "@/components/DayCardHandler.vue";
 // import "swiper/swiper.css";
 
@@ -25,15 +30,7 @@ export default {
     Swiper,
     SwiperSlide,
   },
-  data() {
-    return {
-      swiperOptions: {
-        pagination: {
-          el: ".swiper-pagination",
-        },
-      },
-    };
-  },
+  data() {},
   methods: {
     onSwiper(swiper) {
       console.log(swiper);
@@ -43,15 +40,8 @@ export default {
     },
   },
   props: ["daysData"],
-  computed: {
-    swiper() {
-      return this.$refs.mySwiper.$swiper;
-    },
-  },
-  mounted() {
-    console.log(this.swiper);
-    this.swiper.slideTo(1, 1000, false);
-  },
+  computed: {},
+  mounted() {},
 };
 </script>
 
